Splitted GameControllerDPad class into GameControllerDPadXml and the original
