+++ src/gamecontroller/gamecontrollerdpad.h
@@ -22,11 +22,8 @@
 
 #include "vdpad.h"
 
-class QXmlStreamReader;
 class JoyButton;
 class SetJoystick;
-class JoyDPadXml;
-class DPad;
 
 // holds information about gamecontroller dpad
 
@@ -41,13 +38,7 @@ public:
     virtual QString getName(bool forceFullFormat, bool displayName);
     virtual QString getXmlName(); // GameControllerDPadXml class
 
-    void readJoystickConfig(QXmlStreamReader *xml); // GameControllerDPadXml class
-
     VDPad* getDPad();
-
-private:
-    JoyDPadXml* dpadXml;
-
 };
 
 #endif // GAMECONTROLLERDPAD_H
